requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
filelock==3.6.0
h11==0.13.0
idna==3.3
lxml==4.8.0
outcome==1.1.0
Pillow==9.0.1
pycparser==2.21
pyfiglet==0.8.post1
pyOpenSSL==22.0.0
PySocks==1.6.7
requests==2.27.1
requests-file==1.5.1
selenium==4.1.2
Selenium-Screenshot==1.7.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
stem==1.5.4
tbselenium==0.6.2
tldextract==3.2.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
webdriver-manager==3.5.3
wsproto==1.1.0
